+++ frontend-react/src/hooks/network/Settings/UseOrganizationSettings.test.ts
@@ -1,7 +1,7 @@
 import { renderHook, waitFor } from "@testing-library/react";
 
 import { AppWrapper } from "../../../utils/CustomRenderUtils";
-import { fakeOrg, orgServer } from "../../../__mocks__/OrganizationMockServer";
+import { fakeOrg, settingsServer } from "../../../__mocks__/SettingsMockServer";
 import { mockSessionContext } from "../../../contexts/__mocks__/SessionContext";
 import { MemberType } from "../../UseOktaMemberships";
 import { Organizations } from "../../UseAdminSafeOrganizationName";
@@ -10,10 +10,10 @@ import { useOrganizationSettings } from "./UseOrganizationSettings";
 
 describe("useOrganizationSettings", () => {
     beforeAll(() => {
-        orgServer.listen();
+        settingsServer.listen();
     });
-    afterEach(() => orgServer.resetHandlers());
-    afterAll(() => orgServer.close());
+    afterEach(() => settingsServer.resetHandlers());
+    afterAll(() => settingsServer.close());
     describe("with no Organization name", () => {
         beforeEach(() => {
             mockSessionContext.mockReturnValue({
